+++ tests/sksl/shared/ArrayTypes.metal
@@ -7,7 +7,7 @@ struct S {
 struct Inputs {
 };
 struct Outputs {
-    float4 sk_FragColor [[color(0)]];
+    half4 sk_FragColor [[color(0)]];
 };
 void initialize_vS(thread array<S, 2>& z);
 void _skOutParamHelper0_initialize_vS(thread array<S, 2>& z) {
@@ -30,6 +30,6 @@ fragment Outputs fragmentMain(Inputs _in [[stage_in]], bool _frontFacing [[front
     y[1] = float2(-1.0, 2.0);
     array<S, 2> z;
     _skOutParamHelper0_initialize_vS(z);
-    _out.sk_FragColor = float4(x[0].x * x[0].y + z[0].v.x, x[1].x - x[1].y * z[0].v.y, (y[0].x / y[0].y) / z[1].v.x, y[1].x + y[1].y * z[1].v.y);
+    _out.sk_FragColor = half4(half(x[0].x * x[0].y + z[0].v.x), half(x[1].x - x[1].y * z[0].v.y), half((y[0].x / y[0].y) / z[1].v.x), half(y[1].x + y[1].y * z[1].v.y));
     return _out;
 }
