+++ tests/sksl/shared/RectangleTexture.metal
@@ -4,15 +4,15 @@ using namespace metal;
 struct Inputs {
 };
 struct Outputs {
-    float4 sk_FragColor [[color(0)]];
+    half4 sk_FragColor [[color(0)]];
 };
 struct Globals {
-    texture2d<float> test2D;
+    texture2d<half> test2D;
     sampler test2DSmplr;
-    texture2d<float> test2DRect;
+    texture2d<half> test2DRect;
     sampler test2DRectSmplr;
 };
-fragment Outputs fragmentMain(Inputs _in [[stage_in]], texture2d<float> test2D[[texture(0)]], sampler test2DSmplr[[sampler(0)]], texture2d<float> test2DRect[[texture(1)]], sampler test2DRectSmplr[[sampler(1)]], bool _frontFacing [[front_facing]], float4 _fragCoord [[position]]) {
+fragment Outputs fragmentMain(Inputs _in [[stage_in]], texture2d<half> test2D[[texture(0)]], sampler test2DSmplr[[sampler(0)]], texture2d<half> test2DRect[[texture(1)]], sampler test2DRectSmplr[[sampler(1)]], bool _frontFacing [[front_facing]], float4 _fragCoord [[position]]) {
     Globals _globals{test2D, test2DSmplr, test2DRect, test2DRectSmplr};
     (void)_globals;
     Outputs _out;
