+++ tests/sksl/shared/StaticSwitchWithStaticConditionalBreak.metal
@@ -4,7 +4,7 @@ using namespace metal;
 struct Inputs {
 };
 struct Outputs {
-    float4 sk_FragColor [[color(0)]];
+    half4 sk_FragColor [[color(0)]];
 };
 fragment Outputs fragmentMain(Inputs _in [[stage_in]], bool _frontFacing [[front_facing]], float4 _fragCoord [[position]]) {
     Outputs _out;
@@ -17,6 +17,6 @@ fragment Outputs fragmentMain(Inputs _in [[stage_in]], bool _frontFacing [[front
         case 1:
             x = 1.0;
     }
-    _out.sk_FragColor = float4(x);
+    _out.sk_FragColor = half4(half(x));
     return _out;
 }
