+++ tests/sksl/shared/StructMaxDepth.metal
@@ -54,5 +54,5 @@ struct Inputs {
     array<SA8, 9> sa8;
 };
 struct Outputs {
-    float4 sk_FragColor [[color(0)]];
+    half4 sk_FragColor [[color(0)]];
 };
